add job delete confirm prompt

+++ src/views/JobView.vue
@@ -16,9 +16,12 @@ const state = reactive({
 });
 const deleteJob = async () => {
   try {
-    await axios.delete(`/api/jobs/${jobId}`);
-    toast.success("Job Deleted Successfully");
-    router.push(`/jobs`);
+    const confirm = window.confirm("Are you sure you want to delete this job?");
+    if (confirm) {
+      await axios.delete(`/api/jobs/${jobId}`);
+      toast.success("Job Deleted Successfully");
+      router.push(`/jobs`);
+    }
   } catch (error) {
     console.error("Error deleteing job", error);
     toast.error("Job delete failed");
